backend/requirements.txt
flask
flask-cors
transformers
torch
sentencepiece
gTTS
easyocr
pdf2image
Pillow
accelerate 
optimum